Fixes: Attempts to create folders that already exists and delete nonexistent folders and prints exception #6 by checking if the folder exists first.

+++ disk_tools.py
@@ -34,16 +34,18 @@ def cleanUp(files):
         os.remove(f)
 
 def makeFolder(folder):
-    try:
-        os.mkdir(folder)
-    except OSError as ex:
-        print(ex)
+    if not os.path.isdir(folder):
+        try:
+            os.mkdir(folder)
+        except OSError as ex:
+            print(ex)
 
 def deleteFolder(folder):
-    try:
-        shutil.rmtree(folder)
-    except OSError as ex:
-        print(ex)
+    if os.path.isdir(folder):
+        try:
+            shutil.rmtree(folder)
+        except OSError as ex:
+            print(ex)
 
 if __name__ == '__main__':
     pass
